pyproject.toml
[tool.poetry]
name = "clox"
version = "0.1.0"
description = ""
authors = ["Drew Thomas <cdthomas.92@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
lark-parser = "^0.11.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
